ANDROID: REVERTME: HACK: kvm: pin all the pages used by protected VM
This is a hack implementation to pin every mappings created for a
protected VM. It is probably that one page can be counted multiple
times if KVM is switching EPT root during runtime. All these pinned
pages will be unpinned after this protected VM is destroyed.

For the proper solution, we would like to leverage the file based memory
solution used by TDX[1], which the discussion is still on-going at the
mailing list.

--
[1]
https://lore.kernel.org/linux-mm/CAGtprH_83CEC0U-cBR2FzHsxbwbGn0QJ87WFNOEet8sineOcbQ@mail.gmail.com/

Bug: 395299836
Test: Boot, verify cpus are de-privileged and run a minimal protected vm.

Change-Id: I8da825e1982082b26c501db4dc6f38a842d47584
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>
Signed-off-by: Vineeth Pillai <vineethrp@google.com>