ANDROID: pkvm: x86: add pkvm_pgtable as flush_tlb parameter
The pkvm_pgtable parameter will be used when implementing the shadow EPT
flush_tlb callback in the next patch, to let the callback knows which
EPT tlb should be flushed. This parameter may not be necessary for host
EPT flush_tlb callback as host EPT is unique but shadow EPT is not.

Bug: 395299836
Test: Boot, verify cpus are de-privileged and run a minimal protected vm.

Change-Id: I587f767f59a31a9bce98a94c95cf94e6efbcc7bf
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>
Signed-off-by: Vineeth Pillai <vineethrp@google.com>