ANDROID: pkvm: x86: Dynamically handle host MMIO EPT violation
The host EPT is prepopulated for all the memory and MMIO in the low
address range, but not for the MMIO in high address range. If host VM is
accessing such MMIO, EPT violation happen and pKVM shall do the map
for it if the MMIO range belongs to the host VM, which means this MMIO
cannot be pKVM owned devices(eg. IOMMU) or any pass-through device to
the protected VM (this part will be supported in the future)

Bug: 395299836
Test: Boot, verify cpus are de-privileged and run a minimal protected vm.

Change-Id: Id6086cbb5305df00b9a0f5c9340e9b93bd7fbaca
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>
Signed-off-by: Vineeth Pillai <vineethrp@google.com>