UPSTREAM: soc: mediatek: mtk-svs: Change the thermal sensor device name
This driver tries to create a device link to the thermal sensor device:
change all instances of "lvts" and "thermal" to "thermal-sensor", as
that's what the devicetree node name must be.

Note for MT8183: As specified in a previous commit, this SoC never got
SVS probing, so this is not a breaking change and it does not require
fallback for older device trees.

Link: https://lore.kernel.org/r/20231121125044.78642-7-angelogioacchino.delregno@collabora.com
Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
(cherry picked from commit 30d83ef88feb79a3c9c940c404b30bd351d9f327)

BUG=b:360057331
TEST=build kernel

Change-Id: Ib5f91abe93bd87cdcc6f7c52d47a17459654abb1
Signed-off-by: Hsin-Te Yuan <yuanhsinte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5999844
Commit-Queue: Pin-yen Lin <treapking@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>