UPSTREAM: drm/i915/gem: Atomically invalidate userptr on mmu-notifier
Never block for outstanding work on userptr object upon receipt of a
mmu-notifier. The reason we originally did so was to immediately unbind
the userptr and unpin its pages, but since that has been dropped in
commit b4b9731b02c3c ("drm/i915: Simplify userptr locking"), we never
return the pages to the system i.e. never drop our page->mapcount and so
do not allow the page and CPU PTE to be revoked. Based on this history,
we know we are safe to drop the wait entirely.

Upon return from mmu-notifier, we will still have the userptr pages
pinned preventing the following PTE operation (such as try_to_unmap)
adjusting the vm_area_struct, so it is safe to keep the pages around for
as long as we still have i/o pending.

We do not have any means currently to asynchronously revalidate the
userptr pages, that is always prior to next use.

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Signed-off-by: Jonathan Cavitt <jonathan.cavitt@intel.com>
Reviewed-by: Andi Shyti <andi.shyti@linux.intel.com>
Signed-off-by: Andi Shyti <andi.shyti@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20231128162505.3493942-1-jonathan.cavitt@intel.com
(cherry picked from commit 86ceaaaec59707b06216a15b3852867fa2f1574e)
Signed-off-by: Lee Shawn C <shawn.c.lee@intel.com>

BUG=b:369985652
TEST=No kernel panic while running system stress test.

Change-Id: I0bcc6275dec16b0b5722cad7e2c0190d3724f06f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5951947
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Reviewed-by: Shyam Sundar Radjacoumar <ssradjacoumar@google.com>
Commit-Queue: Shawn C Lee <shawn.c.lee@intel.com>
Tested-by: Wentao Qin <qinwentao@huaqin.corp-partner.google.com>
Signed-off-by: Hubert Mazur <hmazur@google.com>