UPSTREAM: net: wwan: t7xx: Prefer struct_size over open coded arithmetic
This is an effort to get rid of all multiplications from allocation
functions in order to prevent integer overflows [1][2].

As the "port_prox" variable is a pointer to "struct port_proxy" and
this structure ends in a flexible array:

struct port_proxy {
	[...]
	struct t7xx_port ports[];
};

the preferred way in the kernel is to use the struct_size() helper to
do the arithmetic instead of the argument "size + size * count" in the
devm_kzalloc() function.

This way, the code is more readable and safer.

Link: https://www.kernel.org/doc/html/latest/process/deprecated.html#open-coded-arithmetic-in-allocator-arguments [1]
Link: https://github.com/KSPP/linux/issues/160 [2]

Signed-off-by: Erick Archer <erick.archer@gmx.com>
Reviewed-by: Sergey Ryazanov <ryazanov.s.a@gmail.com>
Link: https://lore.kernel.org/r/20240224181932.2720-1-erick.archer@gmx.com
Signed-off-by: Jakub Kicinski <kuba@kernel.org>

(cherry picked from commit 848e34ca203046c9b967034596828472f08e4ac7)
Signed-off-by: Liu Qf <liuqf@fibocom.corp-partner.google.com>

BUG=b:302351783
TEST=check fastboot flash support on brya.

Change-Id: I79328be78b58dadd7561d594310c7ccbd97733ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5928762
Reviewed-by: Ujjwal Pande <ujjwalpande@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Signed-off-by: Hubert Mazur <hmazur@google.com>