UPSTREAM: regulator: devres: add API for reference voltage supplies
A common use case for regulators is to supply a reference voltage to an
analog input or output device. This adds a new devres API to get,
enable, and get the voltage in a single call. This allows eliminating
boilerplate code in drivers that use reference supplies in this way.

Signed-off-by: David Lechner <dlechner@baylibre.com>
Link: https://lore.kernel.org/r/20240429-regulator-get-enable-get-votlage-v2-1-b1f11ab766c1@baylibre.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit b250c20b64290808aa4b5cc6d68819a7ee28237f)

BUG=b:313564065
TEST=cros build-packages -b kukui chromeos-kernel-6_6

Change-Id: I9a18d111631827cfc9834f161477fceb523e4792
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6072619
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Signed-off-by: Hubert Mazur <hmazur@google.com>