UPSTREAM: regulator: core: Fix regulator_is_supported_voltage() kerneldoc return value
The kerneldoc for regulator_is_supported_voltage() states that the
return value is a boolean. That is not correct, as it could return an
error number if the check failed.

Fix the description by expanding it to cover the valid return values and
error conditions. The description is also converted to a proper "Return"
section.

Fixes: c5f3939b8fe0 ("regulator: core: Support fixed voltages in regulator_is_supported_voltage()")
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Reported-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Reviewed-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Link: https://patch.msgid.link/20240829085131.1361701-3-wenst@chromium.org
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 753b9d86adb9d2c5882ac8ee0c3816aa48697eaa)

BUG=b:313564065
TEST=cros build-packages -b kukui chromeos-kernel-6_6

Change-Id: I1013250cecce6176a6ad7b0e37c42fd069cf6b0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6076444
Tested-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Chen-Yu Tsai <wenst@chromium.org>
Signed-off-by: Hubert Mazur <hmazur@google.com>