ANDROID: [TEMP] GKI: x86: Enable PCI for Intel LPSS
This driver supports Intel Low Power Subsystem (LPSS) devices
such as I2C, SPI and HS-UART in PCI mode. It may be required
for some Intel platforms to make serial console available
from the very beginning.

This change is merged as-is into the android-mainline-desktop branch
as part of KCU enablement. All related issues and proposed rework
required for the android-mainline branch are tracked under b/376448213.

Bug: 370694457
Bug: 376442756
Bug: 384007570
Bug: 397385581
Upstream-Task: 376448213
Change-Id: I905c5d93e5b6b606dc39cd0623288cffef73ee40
Signed-off-by: Slawomir Rosek <srosek@google.com>