ANDROID: pkvm: x86: Keep the __pkvm_ prefix for all pkvm symbols in debug mode
The previous implementation is to remove the "__pkvm_" prefix in debug
mode so that can reuse the linux kernel printk symbols. The side effects
of doing this are: 1) not able to detect if pkvm has involved any other
linux kernel symbols or not if develop in debug mode. 2) Not to able
verify the libraries implemented by pkvm as the pkvm's implementation
should be disabled due to the conflict of library symbols. Especially
the 2nd side effact will cause troubles for code sharing.

To resolve these, introduced a debug.c which will be compiled as part
of the linux kernel code (not pkvm hypervisor code) for debug mode only.
As the debug mode also adds the "__pkvm_" prefix to the required linux
kernel printk symbols, the debug.c implements a trampoline for this.

dmaluka: Backport-squashed Zhiyuan's patch "Use macro for debug symbol
extension" into this patch, in order to apply this patch after Zhiyuan's
patch "Change prefix to suffix in patching global symbols" (to prevent
this patch from introducing kCFI regression).

Bug: 391539939
Bug: 391377467
Test: pKVM works
Upstream-Task: 402758258

Change-Id: Id61890cc0a52008d83f1bb81091b079ba8144b9c
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>
Signed-off-by: Dmytro Maluka <dmaluka@google.com>