UPSTREAM: of: dynamic: Add of_changeset_update_prop_string
Add a helper function to add string property updates to an OF changeset.
This is similar to of_changeset_add_prop_string(), but instead of adding
the property (and failing if it exists), it will update the property.

This shall be used later in the DT hardware prober.

Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: Rob Herring (Arm) <robh@kernel.org>
Reviewed-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Signed-off-by: Wolfram Sang <wsa+renesas@sang-engineering.com>
(cherry picked from commit 81de291d86b704de1809cfb06672902d003cf3a3)

BUG=b:313564065
TEST=cros build-packages -b kukui chromeos-kernel-6_6

Change-Id: Id72c1d3163925cbfb8ca16d24a9bb20bd4dcd66e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6076462
Tested-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Signed-off-by: Hubert Mazur <hmazur@google.com>