ANDROID: pkvm: vmx: Add struct pkvm_vm_vmx
Struct pkvm_vm_vmx represents a vm structure for VMX. It requires struct
kvm sitting at offset 0 to be appended in the end of pkvm_vm. As struct
kvm is the first field of struct kvm_vmx, kvm_vmx would also be the
first field of pkvm_vm_vmx.

Bug: 391539939
Test: pKVM works
Upstream-Task: 402758258

Change-Id: I859c2f42466f3fe1462057e1eedf750261cb6dfc
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Dmytro Maluka <dmaluka@google.com>