ANDROID: pkvm: x86: Bump up KVM_X86_PROTECTED_VM to a temp value
In the latest upstream kernel, the set of supported vm_types is:

	#define KVM_X86_DEFAULT_VM	0
	#define KVM_X86_SW_PROTECTED_VM	1
	#define KVM_X86_SEV_VM		2
	#define KVM_X86_SEV_ES_VM	3
	#define KVM_X86_SNP_VM		4

Likely there will be also "#define KVM_X86_TDX_VM 5" in a near future.
We might hope that after that there will be no more vm_types until we
upstream pKVM-IA, but there is no such guarantee.

So for now, pick a "safe" temporary vm_type value that is unlikely to
conflict with future upstream values (but still is below 32, to be
within the 32-bit supported_vm_types mask), to keep the downstream
pKVM-IA uAPI stable for crosvm, but with the intention to replace this
vm_type value with a final value in the future when upstreaming pKVM-IA.

Or possibly in the future KVM_X86_SW_PROTECTED_VM will be repurposed
for pKVM-IA's protected VMs. See commit log in commit 89ea60c2c7b5
("KVM: x86: Add support for "protected VMs" that can utilize private
memory") which introduced KVM_X86_SW_PROTECTED_VM. But we are not there
yet.

Bug: 395299836
Test: Boot, verify cpus are de-privileged and run a minimal protected vm.

Change-Id: I091282fdf3a851a956403bef28d73ec989dbacd2
Signed-off-by: Dmytro Maluka <dmaluka@google.com>
Signed-off-by: Vineeth Pillai <vineethrp@google.com>