UPSTREAM: thunderbolt: Expose router DROM through debugfs
Router DROM contains information that might be usable for development
and debugging purposes. For example when new entries are added to the
USB4 spec it is useful to be able to look for them without need to
change the kernel.

For this reason expose the DROM through debugfs.

(cherry picked from commit 43d84701d2aa147eab39b529919ffaf35f724bbb )
Bug: 401029689
Test: Boot to UI

Change-Id: I806e3497874d7b16e0c9142d6817618bc09be40b
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Fnu Vishwanatha<fnu.vishwanatha@intel.corp-partner.google.com>