ANDROID: arm64: config: Fix the modprobe path
Once aosp/3421379 and aosp/3421399 are merged the modprobe is created
in the correct location the ramdisk and can be safely ran by the kernel.

This change modifies the modprobe path in the kconfig to a expected and
correct location. It's necessary for removing the kvm iommu stub kernel
module from being builtin in to the kernel image.

Bug: 383081956
Test: vm run-microdroid --proctected
Change-Id: I93366d2426ea74ad8cbb81993beee93923da5827
Signed-off-by: Bartłomiej Grzesik <bgrzesik@google.com>