CHROMIUM: arm64: Add command line to limit number of armv8pmu counters used.
The "armv8pmu_max_events=" command line limits the number of pmu events
used with the armv8pmu.

On heterogenous machines where CPUs have different numbers of PMU counters,
the initialization of the PMU might be done on a CPU that has more than
the ones where the PMU counters would get used.
With this command line, the VMM can avoid this situation by setting it
to the number on the smallest CPU.

BUG=b:369260921
TEST=Ran 'simpleperf stat --print-hw-counter' in VM after adding
    "--params=armv8pmu_max_events=20" in /usr/local/vms/etc/arcvm_dev.conf.

Signed-off-by: Suleiman Souhlal <suleiman@google.com>
Change-Id: I3ef9e5cdb609b0816caae26135d457f61d32cfa7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6170758
Reviewed-by: Shao-Chuan Lee <shaochuan@chromium.org>
Reviewed-by: Satoshi Niwa <niwa@chromium.org>
Tested-by: Shao-Chuan Lee <shaochuan@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>