ANDROID: [TEMP] GKI: x86: Convert INT340X and INTEL_SOC_DTS to be GKI modules
When scanning ACPI bus, currently GKI core instantiates platform
devices for INT340X and INTEL_SOC_DTS nodes as long as kernel is compiled
together with INT340X and INTEL_SOC_DTS modules by using IS_ENABLED()
i.e. CONFIG_INT340X_THERMAL=m and CONFIG_INTEL_SOC_DTS_THERMAL=m
in kernel configuration (see int340x_thermal_handler_attach() in
int340x_thermal.c). For GKI, IS_ENABLED() gives positive value
if the config is part of gki_defconfig file which is not the case if
those are vendor modules.

The general problem with IS_ENABLED() was already described here
b/259510298 and the recommendations are to either convert modules
to be GKI modules or refactor the code to avoid IS_ENABLED() usage.
The latter is under investigation but currently there is no viable
(safe for regressions) fix proposed, and until then, for time being,
modules will become GKI modules.

As part of above change, ACPI_FAN is selected, so fan become GKI
module as well.

This change is merged as-is into the android-mainline-desktop branch
as part of KCU enablement. The upstream effort for the android-mainline
branch is tracked under b/394562775.

Bug: 393308727
Bug: 397385581
Upstream-Task: 394562775
Signed-off-by: Tomasz Nowicki <tnowicki@google.com>
Signed-off-by: Mateusz Urbańczyk <urbanczyk@google.com>

Recent changes on the android-mainline branch conflicted with temporary
patches on the andorid-mainline-desktop branch, thus they are reverted
and reapplied after merge. Eventually they are going to be reverted too.
Related patches for the upstream are currently under review.

Bug: 448988819
Link: https://android-review.googlesource.com/c/kernel/common/+/3783531
Link: https://android-review.googlesource.com/c/kernel/common/+/3562019
Link: https://lore.kernel.org/all/20251002113404.3117429-1-srosek@google.com/
Bypass-Upstream-Divergence-Tracking

Change-Id: Ibfaae6c26d36fd71518351e0eef69effbd589337
Signed-off-by: Slawomir Rosek <srosek@google.com>