ANDROID: GKI: set CONFIG_HZ=1000
The current default is 250Hz, which leads to intention of
sleeping for less than 4ms to get an unexpected big delay
that rounds up to 4ms.

This is particularly fatal for low-latency applications such as
audio processing, where 4ms of delay is unacceptable and we
don't want to do busy-wait in such circumstances.

This aligns the frequency with linux for desktop.

Note that this is a temporary solution for android15-6.6-desktop and as
discussed in CL:3518412 and CL:3513330, we will work on more suitable
ways of landing changes like these.

Bug: 395667715
Test: verify audio lag is gone on Brya
Change-Id: I39fbc24de236640ee2bffffb93d004ac44706bad
Signed-off-by: Jeremy Wu <jrwu@google.com>