UPSTREAM: thunderbolt: Do not add non-active NVM if NVM upgrade is disabled for retimer
This is only used to write a new NVM in order to upgrade the retimer
firmware. It does not make sense to expose it if upgrade is disabled.
This also makes it consistent with the router NVM upgrade.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit ad79c278e478ca8c1a3bf8e7a0afba8f862a48a1)

Bug: 427873401
Test: Thunderbolt device is getting enumerated.
Change-Id: I22872c26592cacb7f2b68560024ad069cd408dc1
Signed-off-by: Manvi Bajaj <manvi.bajaj@intel.com>