ANDROID: kvm: pkvm: Use PV interface to do vcpu_after_set_cpuid
Use the PV interface to do the vcpu_after_set_cpuid. Allocating
necessary pages as the cpuid entries memory and send to the pkvm
hypervisor via the PV interface. The pkvm hypervisor will return the
unused page physical address back to the host KVM to free.

Once the guest state is protected, the vcpuid entries are not allowed to
be modified via this PV interface.

Bug: 391539939
Test: pKVM works
Upstream-Task: 402758258

Change-Id: Ic3591311525c16d8a5f789c90b7b52ded1b60e9e
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Dmytro Maluka <dmaluka@google.com>