ANDROID: pkvm: x86: Sync npVM vcpu GPRs between the pkvm hypervisor and the host
As the npVM vcpu doesn't have any vcpu state protection requirement, it
is allowed to keep the vcpu GPRs in sync between the pkvm hypervisor and
the host, to facilitate the vmeixt handling from the both sides.

Introduce the function pkvm_vcpu_update_state_from_host() for the pkvm
hypervisor to update the npVM's vcpu GPRs from the host before entering
the guest as the host may have updated some of the GPRs according to the
vmexit handling result.

Another one pkvm_vcpu_share_state_to_host() is for the pkvm hypervisor to
share the npVM's vcpu GPRs/EFER/CR0/CR4 to the host before returning back
as the host may require to use the update-to-date registers to handle vmexit.

Bug: 391539939
Test: pKVM works
Upstream-Task: 402758258

Change-Id: Ic296c298a20010be0d27ab53e914cb2100795531
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Dmytro Maluka <dmaluka@google.com>