ANDROID: pkvm: x86: Add setup_virtual_mmu ops for pkvm_x86_ops
As currently the EPT solution is still the shadow method, when the host
VMM is doing load_mmu_pgd, the pkvm hypervisor still needs to setup the
virtual mmu for the shadowing purpose. Add a new ops for this. This
should be revisited when the PV EPT method is ready.

Bug: 391539939
Test: pKVM works
Upstream-Task: 402758258

Change-Id: I7ed0450220631ddd9956a0f40166014f427c9fc7
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Dmytro Maluka <dmaluka@google.com>