ANDROID: pkvm: vmx: Share pVM vcpu state to the host for handling vmcall
Some of the vmcall vmexits will be handled by the host KVM, which
requires to access the corresponding guest registers. Share the vcpu
RAX/RBX/RCX/RDX/RSI according to the vmcall protocol before switching
to the host. Meanwhile accept the vmcall handling result in the RAX
shared from the host after switching to the pkvm hypervisor.

Bug: 391539939
Test: pKVM works
Upstream-Task: 402758258

Change-Id: I95287da06199b6d362da6a3bfcf378a5286c6ac5
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Dmytro Maluka <dmaluka@google.com>