ANDROID: [TEMP] GKI: x86: Remove THERMAL_NETLINK from gki_defconfig
Since [1] the THERMAL_NETLINK is selected by INT340X_THERMAL which is
temporarily enabled as a GKI module thus redundant config option needs
to be removed from the gki_defconfig.

This change is merged as-is into the android-mainline-desktop branch
as part of KCU enablement. The upstream effort for the android-mainline
branch is tracked under b/394562775.

[1] https://patch.msgid.link/20241216190821.1137162-1-srinivas.pandruvada@linux.intel.com

Bug: 393308727
Bug: 397385581
Upstream-Task: 394562775
Change-Id: I9fa51acd34ef14612e70bf41eb37e37f57affa86
Signed-off-by: Slawomir Rosek <srosek@google.com>