ANDROID: kvm: pkvm: Add update_cpuid_runtime PV interface
Add update_cpuid_runtime PV interface for the host to ask the pkvm
hypervisor to update the cpuid at the runtime. Currently the cpuid
updating is only happened according to the apic base changes emulated by
the host. So even with this PV interface, the host still cannot
arbitrarily change the cpuid entries for the pVM.

Bug: 391539939
Test: pKVM works
Upstream-Task: 402758258

Change-Id: I30fabd8a477954e6917807d5af68e261a98b3eea
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Dmytro Maluka <dmaluka@google.com>