ANDROID: pkvm: x86: Sync debug registers for the npVM
To support debugging guest from the host side, the host will need to
access the guest debug registers. Meanwhile the pkvm hypervisor handles
the DR_ACCESS vmexit. So the debug registers will be accessed by the
both sides, which requires to keep the debug related registers in sync.
For the security reason, this is only allowed for the npVM.

Bug: 391539939
Test: pKVM works
Upstream-Task: 402758258

Change-Id: Icdf7843e2446f5208e805626d99508cd61276e7d
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Dmytro Maluka <dmaluka@google.com>