ANDROID: xe: throttle attribute cbs take kobject ptrs
sysfs attribute store/show handlers expect the first passed parameter to
be of type 'struct kobject *', but Xe abused the fact that a pointer to
'struct device' can decay to a kobject pointer (its first member).

When CFI is enabled in the kernel, this is detected, resulting in an
intentional kernel crash. Fix the handlers to take a kobject pointer.

Bug: 411102572
Test: adb shell \
    cat /sys/devices/pci0000:00/0000:00:02.0/tile0/gt0/freq0/throttle/*
    doesn't trigger CFI failure, prints reasonable values.
Upstream-Task: 411758923
Change-Id: I803e31c5c1afdac427889ca59c519436b376b958
Signed-off-by: Ryan Neph <ryanneph@google.com>