ANDROID: xe: vram_freq attribute cbs take kobject ptrs
sysfs attribute store/show handlers expect the first passed parameter to
be of type 'struct kobject *', but Xe abused the fact that a pointer to
'struct device' can decay to a kobject pointer (its first member).

When CFI is enabled in the kernel, this is detected, resulting in an
intentional kernel crash. Fix the handlers to take a kobject pointer.

Bug: 411102572
Test: lunch brya-kernel_staging-eng && \
    vendor/google/desktop/dev/kernel/replace_prebuilts.py \
    --build-kernel <kernel_repo_dir>
Upstream-Task: 411758923
Change-Id: I96a9ab2bc30e2753cd42e671895f416dad172100
Signed-off-by: Ryan Neph <ryanneph@google.com>