ANDROID: xe: device sysfs attribute cbs take kobject ptrs
sysfs attribute store/show handlers expect the first passed parameter to
be of type 'struct kobject *', but Xe abused the fact that a pointer to
'struct device' can decay to a kobject pointer (its first member).

When CFI is enabled in the kernel, this is detected, resulting in an
intentional kernel crash. Fix the handlers to take a kobject pointer.

Bug: 411102572
Test: lunch brya-kernel_staging-eng && \
    vendor/google/desktop/dev/kernel/replace_prebuilts.py \
    --build-kernel <kernel_repo_dir>
Upstream-Task: 411758923
Change-Id: I3dbd4c80d8e30ac098d2f9acf4e5b70fe9593810
Signed-off-by: Ryan Neph <ryanneph@google.com>