ANDROID: HID: multitouch: Add lid handler for touchpad
If user closes the lid the touchscreen gets close to the touchpad surface,
which causes interference and makes the touchpad enter noise mode.
Right after opening the lid the cursor is unresponsive because of the mode
mentioned.

To fix this issue we switch the surface off when closing the lid and on
then opening it. This way the FW may perform recalibration once
the lid has been opened.

Bug: 221094337
Bug: 409612576
Test: Build and deploy to device
Upstream-Task: 414567861
Signed-off-by: Sean O'Brien <seobrien@google.com>
Change-Id: I594df06a046c1d9caf6e311df31f59e19c153070