FROMLIST: usb: typec: Implement alternate mode priority handling
This patch introduces APIs to manage the priority of USB Type-C alternate
modes. These APIs allow for setting and retrieving a priority number for
each mode. If a new priority value conflicts with an existing mode's
priority, the priorities of the conflicting mode and all subsequent modes
are automatically incremented to ensure uniqueness.

Signed-off-by: Andrei Kuchynski <akuchynski@chromium.org>

Bug: 382333876
Upstream-Task: 441243047
Test: `/sys/class/typec/portN/portN.M/priority` attribute should be
accessible for both read and write operations.
Link: https://lore.kernel.org/all/20250912064751.1987850-4-akuchynski@chromium.org/

Change-Id: Ie4450416625b11d04bad5b05ddee7dc3f66f61c2
Signed-off-by: Andrei Kuchynski <akuchynski@google.com>