ANDROID: pkvm: x86: Prevent access pkvm_vcpu on a wrong CPU
The load_pkvm_vcpu() will return the pkvm_vcpu via the given handle from
the host for a specific PV interface, and switch from the host vcpu to
the guest. It should prevent the host to use a pkvm_vcpu which is not
loaded on this CPU. The previous implementation can only prevent the
pkvm hypervisor switch to the guest if the pkvm_vcpu is not loaded on
this CPU, but still returns the valid pkvm_vcpu pointer. So add a check
in the load_pkvm_vcpu() to prevent both, so that if a pkvm_vcpu is not
loaded on this CPU, no switch happens and a NULL pointer will return.
The vcpu_load PV interface is an exception as the pkvm_vcpu is not
loaded yet.

Bug: 391539939
Upstream-Task: 402758258

Fixes: cff23c5293fd ("pkvm: x86: Add vcpu_load PV interface")
Reported-by: Dmytro Maluka <dmaluka@google.com>
Suggested-by: Dmytro Maluka <dmaluka@google.com>
Change-Id: I78b05e8ed6e2140dbe4231f71df0c7ee0dc9dd95
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Dmytro Maluka <dmaluka@google.com>