BACKPORT: FROMGIT: tracing: ring_buffer: Rewind persistent ring buffer on reboot
Rewind persistent ring buffer pages which have been read in the previous
boot. Those pages are highly possible to be lost before writing it to the
disk if the previous kernel crashed. In this case, the trace data is kept
on the persistent ring buffer, but it can not be read because its commit
size has been reset after read.  This skips clearing the commit size of
each sub-buffer and recover it after reboot.

Note: If you read the previous boot data via trace_pipe, that is not
accessible in that time. But reboot without clearing (or reusing) the read
data, the read data is recovered again in the next boot.

Thus, when you read the previous boot data, clear it by `echo > trace`.

Cc: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Link: https://lore.kernel.org/174899582116.955054.773265393511190051.stgit@mhiramat.tok.corp.google.com
Signed-off-by: Masami Hiramatsu (Google) <mhiramat@kernel.org>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>
(cherry picked from commit ca296d32ece38b07113bad64e08add75073a0e2b
 git://git.kernel.org/pub/scm/linux/kernel/git/trace/linux-trace.git for-next)

Conflicts:
   kernel/trace/ring_buffer.c

BUG=417075880
TEST=Confirm prev_boot_dump.trace records more then few seconds

Change-Id: Ic73506f189308190117ce2872b56f48e23ae163c
Signed-off-by: Masami Hiramatsu <mhiramat@google.com>