ANDROID: pkvm: vmx: Push vcpu_vmx in the pkvm RSP rather than kvm_vcpu
With pushing vcpu_vmx in the pkvm RSP, the __pkvm_vmexit_entry ASM code
can be extend to restore the pkvm spec ctrl MSR by using the function
vmx_spec_ctrl_restore_host which takes the vcpu_vmx pointer as the input
by the next patch. Although the vcpu_vmx pointer is the same with the
kvm_vcpu pointer due to the structure layout, changing to vcpu_vmx is
easier for reading. This also results in the pkvm_vmexit_main to take the
vcpu_vmx pointer as the input.

Bug: 420681877
Upstream-Task: 402758258

Change-Id: If2f1ae4e8d6169aa1e9732400862cf40c0ab383f
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>