FROMLIST: drm: Add Color ops capability property
Add capability property which a colorop can expose it's
hardware's abilities. It's a blob property that can be
filled with respective data structures depending on the
colorop. The user space is expected to read this property
and program the colorop accordingly.

v2: Added documentation for hw_caps blob (Dmitry)

Signed-off-by: Uma Shankar <uma.shankar@intel.com>
Signed-off-by: Chaitanya Kumar Borah <chaitanya.kumar.borah@intel.com>

Upstream-Task: b:430302249
Bug: 432032023
Test: None
Link: https://lore.kernel.org/r/20250702091936.3004854-4-uma.shankar@intel.com

Change-Id: I9d957f3a4affc2a9ed17319b1267c859cfb93fe7
Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>