ANDROID: pkvm: x86: handle memcache refill during page sharing
When the protected guest shares the pages with host additional mapping
may be required due to splitting a huge page mapping into smaller ones.
In such case there may be not enough memory in memcache which is needed
for extending guest page tables. If it happens return to host and refill
the memcache after which re-issue guest page sharing operation.

Bug: 391541180
Bug: 453743944
Test: Build, boot and run various VMs
Upstream-Task: 402758258

Change-Id: I27a222b25feb255251218a13a3c19c149314997f
Signed-off-by: Grzegorz Jaszczyk <jaszczyk@google.com>