ANDROID: DEBUG: pkvm: x86: Use per vcpu vmexit_perf to trace guest vmexit
The pkvm traces the guest vmexit via calling the trace_vmexit_start
after vmexit occurs, and calling the trace_vmexit_end before vmenter so
that the vmexit reason and cycles will be saved. And the current pkvm
trace is using a global percpu structure to save these trace data for
the guest.

With PV-based VMCS, some of the guest vmexits are handled inside the
pkvm hypervisor only while some are handled cooperate by the host and
the pkvm hypervisor. For such cooperate handled vmexit, the pkvm
hypervisor can switch to the host after calling trace_vmexit_start but
without calling trace_vmexit_end, and the host can schedule another
vcpu to run, which will result in the pkvm calling trace_vmexit_end for
a different vcpu and result in the incorrect trace record.

Fix this via introducing a per vcpu vmexit_perf to avoid such conflict.

Upstream-Task: 402758258

Fixes: c54a3b1bcd04 ("ANDROID: DEBUG: pkvm: x86: add vmexit trace support")
Change-Id: I5eaad42574875be42e44d5fad6c3e0de22a0a0f6
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>