ANDROID: pkvm: x86: Passthrough the PCI configuration space
The pkvm monitors the PCI configuration space for the purpose of
supporting passthrough device to the pVM, which can prevent the host to
change the BAR in the passthrough device's pci configuration space.

However, there is no requirement to passthrough devices to pVM yet, so
this monitoring is only causing unneeded extra complexity and overhead
for now. Moreover, it is also causing various runtime issues (for
instance, issues with Thunderbolt on PTL, or issues with ACPI AML code
accessing PCI config space on some platforms), likely caused by the fact
that for monitoring PCI MMCFG access it uses a very limited MMIO
instruction decoder which doesn't work in some cases.

Passthrough the PCI configuration space to the host by removing the
code which monitors the PCI PIO and ECAM MMIO. As well as the io_bitmap
and io emulation support.

Bug: 434032928
Upstream-Task: 402758258

Change-Id: I69d8961f49efa86a16e1c78bf0c659cf89b99b2f
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>