ANDROID: ASoC: Intel: sof_sdw: use PRODUCT_FAMILY for Fatcat series
PRODUCT_NAME is machine-specific. Use PRODUCT_FAMILY to ensure
the machine quirk is applied with consistent audio configurations
across Fatcat series products.

Although it's tagged with ANDROID, this patch is already uploaded to sof GitHub. And usually commits to sof GitHub will be posted to the Linux kernel mailing list in batch.

Bug: 431945678
Upstream-Task: 441399259
Test: verify the sound card list
Link: https://github.com/thesofproject/linux/pull/5520/commits/

Change-Id: I65ce6d85b108fd8ea4d3ef8def4c2cb71b2fa60c
Signed-off-by: Mac Chiang <mac.chiang@intel.com>