FROMLIST: platform/chrome: cros_usbpd_notify: defer probe when parent EC driver isn't ready
The cros-usbpd-notify-acpi probe currently does not exit when it fails
to get a pointer to the ChromeOS EC device. It is expected behavior on
older devices, where GOOG0004 is not a parent of GOOG0003.

Update the cros-usbpd-notify-acpi probe to check for a GOOG0004 parent
fwnode. If the device has correct device hierarchy and fails to get an
EC device pointer, defer the probe function.

Reviewed-by: Benson Leung <bleung@chromium.org>
(am from https://patchwork.kernel.org/patch/14248233/)
(also found at https://lore.kernel.org/r/20250926202234.3064386-1-jthies@google.com)

Bug: 437116596
Upstream-Task: 448166179
Test: cros_ec_ucsi probe succeeds

Link: https://patchwork.kernel.org/patch/14248233
Link: https://lore.kernel.org/r/20250926202234.3064386-1-jthies@google.com
Change-Id: I51a4eff6d97d24056eb139754fb5340711f58bfa
Signed-off-by: Jameson Thies <jthies@google.com>