ANDROID: pkvm: x86: Remove FIXME about kvm_mmu_after_set_cpuid()
Either with or without pKVM, kvm_mmu_after_set_cpuid() should be called
after changing the guest cpuid, to let the host update the guest MMU
info that depends on the cpuid (for instance, mmu->guest_rsvd_check
which is used when walking guest page tables for instruction emulation
for npVM).

However, luckily, we don't need to do that in pKVM hypervisor, since it
is done in the host anyway, in kvm_vcpu_after_set_cpuid() after
returning from the vcpu_after_set_cpuid PV call to pKVM (and this PV
call is the only code path in pKVM hypervisor where we change the guest
cpuid).

Bug: 391541180
Test: run various VMs
Upstream-Task: 402758258

Change-Id: I6647278372da762e2cd94bddf724e57ed87ea65e
Signed-off-by: Dmytro Maluka <dmaluka@google.com>