ANDROID: pkvm: x86: Deprecate PKVM_GHC_GET_VE_INFO
This hypercall is neither implemented by the up-to-date pKVM hypervisor
code (in 6.12+ branches) nor used by up-to-date pKVM guests, so
deprecate it.

The actual motivation for this cleanup is to free up the hypercall
number used by PKVM_GHC_GET_VE_INFO to use it for the upcoming new
hypercall for starting secondary CPUs.

As the pKVM hypervisor code in android15-6.6-desktop is out of date
and still uses/implements PKVM_GHC_GET_VE_INFO, still do not remove its
definition, so that compiling android15-6.6-desktop as a host kernel
with CONFIG_PKVM_INTEL=y still succeeds. Note, however, that
android15-6.6-desktop is not used as a host kernel on x86 in real
setups anymore, it is only used as a guest kernel.

Bug: 374338925
Test: run various VMs
Upstream-Task: 402758258

Change-Id: Icb073be3eb37139b412b31d3d19473cd9db69f24
Signed-off-by: Dmytro Maluka <dmaluka@google.com>