ANDROID: pkvm: x86: Allow PKVM_INTEL and PKVM_GUEST at the same time
When PKVM_GUEST is enabled, during runtime the kernel detects if the
system is a pkvm guest or not, in order to change some behaviour.

If we want to use the same kernel for host and guest we need to enable
PKVM_GUEST and PKVM_INTEL at the same time.

Remove this artificial limitation.

Test: tools/bazel run //common:kernel_x86_64_dist
Bug: 444310763
Upstream-Task: 402758258

Change-Id: Ie0f1c3eb6cad1388c034f1f973c14a6f37d2fc92
Signed-off-by: Ricardo Ribalda <ribalda@google.com>