ANDROID: ALSA: sound: Increase snd_card components length
The components field of snd_card can run out of space in new systems
which use many audio devices, hence increase its size to 256 bytes.

It's a partial change of https://patchwork.kernel.org/patch/14280930.
The whole change can't be submitted because it is causing tinymix
failure. And this change is needed to make lapis soundcard listed.

Bug: 444122406
Bug: 458418448
Upstream-Task: 457971634
Change-Id: I4b2402a57b3dc3f3f53f495df5dd436174783bba
Signed-off-by: Norman Bintang <normanbt@google.com>