FROMLIST: rust_binder: avoid mem::take on delivered_deaths
Similar to the previous commit, List::remove is used on
delivered_deaths, so do not use mem::take on it as that may result in
violations of the List::remove safety requirements.

I don't think this particular case can be triggered because it requires
fd close to run in parallel with an ioctl on the same fd. But let's not
tempt fate.

Cc: stable@vger.kernel.org
Fixes: eafedbc7c050 ("rust_binder: add Rust Binder driver")
Signed-off-by: Alice Ryhl <aliceryhl@google.com>

Bug: 457535597
Link: https://lore.kernel.org/r/20251111-binder-fix-list-remove-v1-2-8ed14a0da63d@google.com
Change-Id: I14ed7a6fe31241af36be825a24c036e0e18083a9
Signed-off-by: Alice Ryhl <aliceryhl@google.com>