FROMLIST: rust: list: add warning to List::remove docs about mem::take
The previous patches in this series illustrate why the List::remove
method is really dangerous. I think the real takeaway here is to replace
the linked lists with a different data structure without this unsafe
footgun, but for now we fix the bugs and add a warning to the docs.

Signed-off-by: Alice Ryhl <aliceryhl@google.com>

Bug: 457535597
Link: https://lore.kernel.org/r/20251111-binder-fix-list-remove-v1-3-8ed14a0da63d@google.com
Change-Id: I35e1dbc3da6ea2a516ca2f565c108ac971ea5871
Signed-off-by: Alice Ryhl <aliceryhl@google.com>