ANDROID: pkvm: x86: Reset vCPU when creating it
Currently pKVM doesn't normally reset the vCPU via kvm_vcpu_reset() when
creating it, it just relies on the host to request that via the
__pkvm__vcpu_reset hypercall, and allows the host to not do that.
So a vCPU may start running in a state that was never properly reset,
only zero-initialized (although that is almost the same thing, but not
quite the same), which is potentially problematic for security.

So let pKVM reset vCPU when creating it, like regular KVM does.

Also avoid calling the __pkvm__vcpu_reset hypercall in the host for
resetting the vCPU after it has been already created via the
__pkvm__vcpu_create hypercall (and thus already reset by pKVM as a
part of that), since this extra reset is now redundant, and to avoid
triggering the WARN_ON_ONCE in kvm_vcpu_reset() which assumes that a
vCPU reset should only be done during vCPU creation, when it is still
fully zero-initialized.

Still use the __pkvm__vcpu_reset hypercall for INIT when starting
secondary vCPUs via INIT-SIPI emulated by the host, as we are still
relying on that at least for non-protected VMs.

Bug: 429123537
Test: run pVMs
Upstream-Task: 402758258

Change-Id: I4e0d75183c46e8ce8368bc2d03d26de9330e5057
Signed-off-by: Dmytro Maluka <dmaluka@google.com>