ANDROID: iommu/arm-smmu-v3-kvm: Fix visibility of global symbols
kvm_smmu_v3_ops is only referenced in arm-smmu-v3-kvm.c, so reduce
its scope to file static global. This fixes the following sparse
warning:

drivers/iommu/arm/arm-smmu-v3/arm-smmu-v3-kvm.c:1510:25: warning: symbol
'kvm_smmu_v3_ops' was not declared. Should it be static?

Similarly, smmu_[suspend|resume] are not referred to outside of
arm-smmu-v3.c in the pkvm subdirectory, so declare them as static.

smmu_ops is actually referenced outside of the file it is declared
in, so declare it as extern to fix the following sparse warning:

drivers/iommu/arm/arm-smmu-v3/pkvm/arm-smmu-v3.c:1808:22: warning: symbol
'smmu_ops' was not declared. Should it be static?

Bug: 461679166
Change-Id: I797697d0e4e9c46ea05efcaf1f5937600a7d9e0d
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>